HW-08

+++ src/redux/contacts/contactsOps.js
@@ -3,9 +3,26 @@ import axios from 'axios';
 
 axios.defaults.baseURL = 'https://connections-api.herokuapp.com';
 
+const setAuthToken = token => {
+  if (token) {
+    axios.defaults.headers.common['Authorization'] = `Bearer ${token}`;
+  } else {
+    delete axios.defaults.headers.common['Authorization'];
+  }
+};
+
 export const fetchContacts = createAsyncThunk(
   'contacts/fetchAll',
   async (_, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue('No token found');
+    }
+
+    setAuthToken(token);
+
     try {
       const response = await axios.get('/contacts');
       return response.data;
@@ -18,6 +35,15 @@ export const fetchContacts = createAsyncThunk(
 export const addContact = createAsyncThunk(
   'contacts/addContact',
   async (contact, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue('No token found');
+    }
+
+    setAuthToken(token);
+
     try {
       const response = await axios.post('/contacts', contact);
       return response.data;
@@ -30,6 +56,15 @@ export const addContact = createAsyncThunk(
 export const deleteContact = createAsyncThunk(
   'contacts/deleteContact',
   async (contactId, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue('No token found');
+    }
+
+    setAuthToken(token);
+
     try {
       await axios.delete(`/contacts/${contactId}`);
       return contactId;
